+++ src/main/php/inject/Injector.class.php
@@ -2,9 +2,11 @@
 
 use lang\Type;
 use lang\XPClass;
-use lang\Generic;
+use lang\TypeUnion;
+use lang\Primitive;
 use lang\Throwable;
 use lang\IllegalArgumentException;
+use lang\IllegalAccessException;
 use lang\mirrors\TypeMirror;
 use lang\mirrors\TargetInvocationException;
 
@@ -27,7 +29,7 @@ class Injector extends \lang\Object {
    * @param  inject.Bindings... $initial
    */
   public function __construct() {
-    $this->bind($this->getClass(), $this);
+    $this->bind(typeof($this), $this);
     foreach (func_get_args() as $bindings) {
       $this->add($bindings);
     }
@@ -36,18 +38,20 @@ class Injector extends \lang\Object {
   /**
    * Returns a binding
    *
-   * @param  lang.XPClass $t
+   * @param  lang.Type $t
    * @param  var $impl
    */
-  protected function asBinding($t, $impl) {
+  public static function asBinding($t, $impl) {
     if ($impl instanceof XPClass) {
-      return new ClassBinding($t, $impl);
+      return new ClassBinding($impl, $t);
     } else if (self::$PROVIDER->isInstance($impl) || $impl instanceof Provider) {
-      return new ProviderBinding($t, $impl);
-    } else if ($impl instanceof Generic) {
-      return new InstanceBinding($t, $impl);
+      return new ProviderBinding($impl);
+    } else if (is_object($impl)) {
+      return new InstanceBinding($impl, $t);
+    } else if (is_array($impl)) {
+      return new ArrayBinding($impl, $t);
     } else {
-      return new ClassBinding($t, XPClass::forName((string)$impl));
+      return new ClassBinding(XPClass::forName((string)$impl), $t);
     }
   }
 
@@ -74,65 +78,47 @@ class Injector extends \lang\Object {
   public function bind($type, $impl, $name= null) {
     $t= $type instanceof Type ? $type : Type::forName($type);
 
-    if ($t instanceof XPClass) {
-      $this->bindings[$t->literal().$name]= $this->asBinding($t, $impl);
-    } else if (null === $name) {
-      throw new IllegalArgumentException('Cannot bind non-class type '.$t.' without a name');
+    if ($impl instanceof Named) {
+      $this->bindings[$t->literal()]= $impl;
+    } else if ($t instanceof Primitive) {
+      if (null === $name) {
+        throw new IllegalArgumentException('Cannot bind primitive type '.$t.' without a name');
+      }
+      $this->bindings[$t->literal()][$name]= new InstanceBinding($impl, $t);
     } else {
-      $this->bindings[$t->literal().$name]= new InstanceBinding($t, $impl);
+      $this->bindings[$t->literal()][$name]= self::asBinding($t, $impl);
     }
-
     return $this;
   }
 
   /**
    * Get a binding
    *
-   * @param  var $type either a lang.Type instance or a type name
+   * @param  string|lang.Type $type
    * @param  string $name
    * @return var or NULL if none exists
    */
   public function get($type, $name= null) {
     $t= $type instanceof Type ? $type : Type::forName($type);
 
-    if (self::$PROVIDER->isAssignableFrom($t)) {
-      if (isset($this->bindings[$combined= $t->genericArguments()[0]->literal().$name])) {
-        return $this->bindings[$combined]->provider($this);
+    if ($t instanceof TypeUnion) {
+      foreach ($t->types() as $type) {
+        if ($instance= $this->get($type, $name)) return $instance;
+      }
+    } else if (self::$PROVIDER->isAssignableFrom($t)) {
+      $literal= $t->genericArguments()[0]->literal();
+      if (isset($this->bindings[$literal][$name])) {
+        return $this->bindings[$literal][$name]->provider($this);
       }
     } else {
-      if (isset($this->bindings[$combined= $t->literal().$name])) {
-        return $this->bindings[$combined]->resolve($this);
-      } else if ($t instanceof XPClass && !($t->isInterface() || $t->getModifiers() & MODIFIER_ABSTRACT)) {
+      $literal= $t->literal();
+      if (isset($this->bindings[$literal][$name])) {
+        return $this->bindings[$literal][$name]->resolve($this);
+      } else if (null === $name && $t instanceof XPClass && !($t->isInterface() || $t->getModifiers() & MODIFIER_ABSTRACT)) {
         return $this->newInstance($t);
       }
     }
     return null;
-  }
-
-  /**
-   * Retrieve bound value for injection
-   *
-   * @param  var $inject The annotation
-   * @param  lang.reflect.Field $field
-   * @return var
-   * @throws inject.ProvisionException
-   */
-  private function field($inject, $field) {
-    $binding= $this->get(
-      isset($inject['type']) ? $inject['type'] : $field->type(),
-      isset($inject['name']) ? $inject['name'] : null
-    );
-
-    if (null === $binding) {
-      throw new ProvisionException(sprintf(
-        'Unknown injection type %s%s for %s\'s field %s',
-        $type,
-        isset($inject['name']) ? ' named "'.$inject['name'].'"' : '',
-        $field->declaredIn()->name(),
-        $field->name()
-      ));
-    }
-    return $binding;
   }
 
   /**
@@ -156,7 +142,7 @@ class Injector extends \lang\Object {
       } else {
         throw new ProvisionException(sprintf(
           'Unknown injection type %s%s for %s\'s %s() parameter %s',
-          $type,
+          isset($inject['type']) ? $inject['type'] : $param->getTypeName(),
           isset($inject['name']) ? ' named "'.$inject['name'].'"' : '',
           $routine->declaredIn()->name(),
           $routine->name(),
@@ -172,11 +158,10 @@ class Injector extends \lang\Object {
    *
    * @param  lang.mirrors.Routine $routine
    * @param  [:var] $named Named arguments
-   * @param  bool $target
    * @return var
    * @throws inject.ProvisionException
    */
-  protected function args($routine, $named, $target) {
+  protected function args($routine, $named) {
     $inject= $routine->annotations()->present('inject');
     $args= [];
     foreach ($routine->parameters() as $i => $param) {
@@ -189,7 +174,7 @@ class Injector extends \lang\Object {
       } else if ($inject) {
         $target= true;
         $args[]= $this->param(0 === $i ? $routine->annotations()->named('inject')->value() : [], $routine, $param);
-      } else if ($target && !$param->isOptional()) {
+      } else if (!$param->isOptional()) {
         throw new ProvisionException(sprintf(
           'Value required for %s\'s %s() parameter %s',
           $routine->getDeclaringClass()->getName(),
@@ -198,7 +183,7 @@ class Injector extends \lang\Object {
         ));
       }
     }
-    return $target ? $args : null;
+    return $args;
   }
 
   /**
@@ -208,52 +193,27 @@ class Injector extends \lang\Object {
    *
    * @param   lang.XPClass $class
    * @param   [:var] $named Named arguments
-   * @return  lang.Generic
+   * @return  var
    * @throws  inject.ProvisionException
    */
   public function newInstance(XPClass $class, $named= []) {
     $mirror= new TypeMirror($class);
     $constructor= $mirror->constructor();
+    $modifiers= $constructor->modifiers();
+
+    if (!$modifiers->isPublic()) {
+      throw new ProvisionException(
+        'Error creating an instance of '.$mirror->name(),
+        new IllegalAccessException('Cannot invoke '.$modifiers->names().' constructor')
+      );
+    }
 
     try {
-      return $constructor->newInstance(...$this->args($constructor, $named, true));
+      return $constructor->newInstance(...$this->args($constructor, $named));
     } catch (TargetInvocationException $e) {
       throw new ProvisionException('Error creating an instance of '.$mirror->name(), $e->getCause());
     } catch (Throwable $e) {
       throw new ProvisionException('Error creating an instance of '.$mirror->name(), $e);
     }
-  }
-
-  /**
-   * Inject members of an instance
-   *
-   * @param   lang.Generic $instance
-   * @return  lang.Generic
-   * @throws  inject.ProvisionException
-   */
-  public function into(Generic $instance) {
-    $mirror= new TypeMirror($instance->getClass());
-
-    foreach ($mirror->fields() as $field) {
-      if (!$field->annotations()->present('inject')) continue;
-      try {
-        $field->modify($instance, $this->field($field->annotations()->named('inject')->value(), $field));
-      } catch (Throwable $e) {
-        throw new ProvisionException('Error setting '.$mirror->name().'::$'.$field->name());
-      }
-    }
-
-    foreach ($mirror->methods() as $method) {
-      if (null === ($args= $this->args($method, [], false))) continue;
-      try {
-        $method->invoke($instance, $args);
-      } catch (TargetInvocationException $e) {
-        throw new ProvisionException('Error invoking '.$mirror->name().'::'.$method->name(), $e->getCause());
-      } catch (Throwable $e) {
-        throw new ProvisionException('Error invoking '.$mirror->name().'::'.$method->name(), $e);
-      }
-    }
-
-    return $instance;
   }
 }