+++ src/main/php/inject/InstanceBinding.class.php
@@ -4,21 +4,19 @@ use lang\IllegalArgumentException;
 use util\Objects;
 
 class InstanceBinding extends \lang\Object implements Binding {
-  protected $type;
   protected $instance;
 
   /**
    * Creates a new instance binding
    *
-   * @param  lang.Type $type
    * @param  var $instance
+   * @param  lang.Type $type
    * @throws lang.IllegalArgumentException
    */
-  public function __construct($type, $instance) {
-    if (!$type->isInstance($instance)) {
+  public function __construct($instance, $type= null) {
+    if ($type && !$type->isInstance($instance)) {
       throw new IllegalArgumentException(Objects::stringOf($instance).' is not an instance of '.$type);
     }
-    $this->type= $type;
     $this->instance= $instance;
   }