+++ src/test/php/inject/unittest/InjectorTest.class.php
@@ -8,20 +8,32 @@ use lang\ClassNotFoundException;
 use unittest\TestCase;
 use util\Currency;
 use inject\unittest\fixture\FileSystem;
+use inject\unittest\fixture\InMemory;
+use inject\unittest\fixture\Storage;
+use inject\unittest\fixture\AbstractStorage;
 
 class InjectorTest extends TestCase {
 
   /** @return var[][] */
   protected function bindings() {
     $instance= new FileSystem();
-    $name= 'inject.unittest.fixture.Storage';
+    $name= Storage::class;
     return [
-      [$name, XPClass::forName('inject.unittest.fixture.FileSystem')],
-      [$name, 'inject.unittest.fixture.FileSystem'],
+      [$name, XPClass::forName(FileSystem::class)],
+      [$name, FileSystem::class],
       [$name, $instance],
-      [XPClass::forName($name), XPClass::forName('inject.unittest.fixture.FileSystem')],
-      [XPClass::forName($name), 'inject.unittest.fixture.FileSystem'],
+      [XPClass::forName($name), XPClass::forName(FileSystem::class)],
+      [XPClass::forName($name), FileSystem::class],
       [XPClass::forName($name), $instance]
+    ];
+  }
+
+  /** @return var[][] */
+  protected function storages() {
+    return [
+      [[XPClass::forName(FileSystem::class), XPClass::forName(InMemory::class)]],
+      [[FileSystem::class, InMemory::class]],
+      [[new FileSystem(), new InMemory()]]
     ];
   }
 
@@ -45,31 +57,38 @@ class InjectorTest extends TestCase {
   #[@test]
   public function binds_self_per_default() {
     $inject= new Injector();
-    $this->assertEquals($inject, $inject->get('inject.Injector'));
+    $this->assertEquals($inject, $inject->get(Injector::class));
   }
 
   #[@test, @values('bindings')]
   public function get_implementation_bound_to_interface($type, $impl) {
     $inject= new Injector();
     $inject->bind($type, $impl);
-    $this->assertInstanceOf('inject.unittest.fixture.FileSystem', $inject->get($type));
+    $this->assertInstanceOf(FileSystem::class, $inject->get($type));
+  }
+
+  #[@test, @values('storages')]
+  public function bind_array($impl) {
+    $inject= new Injector();
+    $inject->bind('inject.unittest.fixture.Storage[]', $impl);
+    $this->assertEquals([new FileSystem(), new InMemory()], $inject->get('inject.unittest.fixture.Storage[]'));
   }
 
   #[@test]
   public function creates_implicit_binding_when_no_explicit_binding_exists_and_type_is_concrete() {
     $inject= new Injector();
-    $impl= 'inject.unittest.fixture.FileSystem';
+    $impl= FileSystem::class;
     $this->assertInstanceOf($impl, $inject->get($impl));
   }
 
   #[@test]
   public function no_implicit_binding_for_interfaces() {
-    $this->assertNull((new Injector())->get('inject.unittest.fixture.Storage'));
+    $this->assertNull((new Injector())->get(Storage::class));
   }
 
   #[@test]
   public function no_implicit_binding_for_abstract_classes() {
-    $this->assertNull((new Injector())->get('inject.unittest.fixture.AbstractStorage'));
+    $this->assertNull((new Injector())->get(AbstractStorage::class));
   }
 
   #[@test]
@@ -101,38 +120,50 @@ class InjectorTest extends TestCase {
   #[@test, @expect(IllegalArgumentException::class)]
   public function cannot_bind_non_concrete_implementation() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', 'inject.unittest.fixture.AbstractStorage');
+    $inject->bind(Storage::class, AbstractStorage::class);
   }
 
   #[@test, @expect(IllegalArgumentException::class)]
   public function cannot_bind_uncompatible_instance() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', $this);
+    $inject->bind(Storage::class, $this);
   }
 
   #[@test, @expect(IllegalArgumentException::class)]
   public function cannot_bind_uncompatible_class() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', XPClass::forName('unittest.TestCase'));
+    $inject->bind(Storage::class, XPClass::forName(TestCase::class));
   }
 
   #[@test, @expect(IllegalArgumentException::class)]
   public function cannot_bind_uncompatible_class_name() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', 'unittest.TestCase');
+    $inject->bind(Storage::class, TestCase::class);
   }
 
   #[@test, @expect(ClassNotFoundException::class)]
   public function cannot_bind_non_existant_class() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', '@non.existant.class@');
+    $inject->bind(Storage::class, '@non.existant.class@');
+  }
+
+  #[@test, @expect(IllegalArgumentException::class)]
+  public function cannot_bind_array_type_to_non_array() {
+    $inject= new Injector();
+    $inject->bind('inject.unittest.fixture.Storage[]', Storage::class);
+  }
+
+  #[@test, @expect(IllegalArgumentException::class)]
+  public function cannot_bind_non_array_type_to_array() {
+    $inject= new Injector();
+    $inject->bind('inject.unittest.fixture.Storage', [Storage::class]);
   }
 
   #[@test, @values('bindings')]
   public function get_named_implementation_bound_to_interface($type, $impl) {
     $inject= new Injector();
     $inject->bind($type, $impl, 'test');
-    $this->assertInstanceOf('inject.unittest.fixture.FileSystem', $inject->get($type, 'test'));
+    $this->assertInstanceOf(FileSystem::class, $inject->get($type, 'test'));
   }
 
   #[@test, @values('bindings')]
@@ -146,5 +177,30 @@ class InjectorTest extends TestCase {
     $inject= new Injector();
     $inject->bind($type, $impl, 'test');
     $this->assertNull($inject->get($type, 'another-name-than-the-one-bound'));
+  }
+
+  #[@test]
+  public function get_given_a_typeunion_searches_all_types() {
+    $fs= new FileSystem('/usr');
+    $inject= new Injector();
+    $inject->bind(FileSystem::class, $fs);
+    $this->assertEquals($fs, $inject->get('string|inject.unittest.fixture.FileSystem'));
+  }
+
+  #[@test]
+  public function get_given_a_typeunion_searches_all_named_types() {
+    $path= '/usr';
+    $inject= new Injector();
+    $inject->bind('string', $path, 'path');
+    $this->assertEquals($path, $inject->get('string|inject.unittest.fixture.FileSystem', 'path'));
+  }
+
+  #[@test]
+  public function get_given_a_typeunion_searches_all_named_types_and_uses_first() {
+    $path= '/usr';
+    $inject= new Injector();
+    $inject->bind('string', $path, 'path');
+    $inject->bind(FileSystem::class, new FileSystem('/usr'));
+    $this->assertEquals($path, $inject->get('string|inject.unittest.fixture.FileSystem', 'path'));
   }
 }