+++ src/test/php/inject/unittest/ProvidersTest.class.php
@@ -1,17 +1,19 @@
 <?php namespace inject\unittest;
 
 use inject\Injector;
+use inject\InstanceProvider;
 use inject\TypeProvider;
 use unittest\TestCase;
 use lang\XPClass;
 use inject\unittest\fixture\FileSystem;
+use inject\unittest\fixture\Storage;
 
 class ProvidersTest extends TestCase {
 
   #[@test]
   public function type_provider() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', XPClass::forName('inject.unittest.fixture.FileSystem'));
+    $inject->bind(Storage::class, XPClass::forName(FileSystem::class));
     $this->assertInstanceOf(
       'inject.TypeProvider',
       $inject->get('inject.Provider<inject.unittest.fixture.Storage>')
@@ -21,9 +23,9 @@ class ProvidersTest extends TestCase {
   #[@test]
   public function type_provider_get() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', XPClass::forName('inject.unittest.fixture.FileSystem'));
+    $inject->bind(Storage::class, XPClass::forName(FileSystem::class));
     $this->assertInstanceOf(
-      'inject.unittest.fixture.FileSystem',
+      FileSystem::class,
       $inject->get('inject.Provider<inject.unittest.fixture.Storage>')->get()
     );
   }
@@ -31,28 +33,28 @@ class ProvidersTest extends TestCase {
   #[@test]
   public function type_provider_bound_to_type_provider() {
     $inject= new Injector();
-    $provider= new TypeProvider(XPClass::forName('inject.unittest.fixture.FileSystem'), $inject);
-    $inject->bind('inject.unittest.fixture.Storage', $provider);
+    $provider= new TypeProvider(XPClass::forName(FileSystem::class), $inject);
+    $inject->bind(Storage::class, $provider);
     $this->assertEquals($provider, $inject->get('inject.Provider<inject.unittest.fixture.Storage>'));
   }
 
   #[@test]
   public function type_bound_to_type_provider() {
     $inject= new Injector();
-    $provider= new TypeProvider(XPClass::forName('inject.unittest.fixture.FileSystem'), $inject);
-    $inject->bind('inject.unittest.fixture.Storage', $provider);
+    $provider= new TypeProvider(XPClass::forName(FileSystem::class), $inject);
+    $inject->bind(Storage::class, $provider);
     $this->assertInstanceOf(
-      'inject.unittest.fixture.FileSystem',
-      $inject->get('inject.unittest.fixture.Storage')
+      FileSystem::class,
+      $inject->get(Storage::class)
     );
   }
 
   #[@test]
   public function instance_provider() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', new FileSystem());
+    $inject->bind(Storage::class, new FileSystem());
     $this->assertInstanceOf(
-      'inject.InstanceProvider',
+      InstanceProvider::class,
       $inject->get('inject.Provider<inject.unittest.fixture.Storage>')
     );
   }
@@ -60,9 +62,9 @@ class ProvidersTest extends TestCase {
   #[@test]
   public function instance_provider_get() {
     $inject= new Injector();
-    $inject->bind('inject.unittest.fixture.Storage', new FileSystem());
+    $inject->bind(Storage::class, new FileSystem());
     $this->assertInstanceOf(
-      'inject.unittest.fixture.FileSystem',
+      FileSystem::class,
       $inject->get('inject.Provider<inject.unittest.fixture.Storage>')->get()
     );
   }